cleanup installer code use int type as index for all installer info arrays correct some another type mismatchs use size_t type where appropriate for better portability

+++ bld/setupgui/h/dynarray.h
@@ -31,7 +31,7 @@
 
 
 typedef struct array_info {
-    size_t      num;
+    int         num;
     size_t      alloc;
     size_t      esize;
     void        **array;
